+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/ShowAboutCommand.cs
@@ -1,0 +1,16 @@
+using System.Windows;
+using TD.WPF.Dialogs;
+
+namespace TD.WPF.Commands;
+
+internal class ShowAboutCommand : AbstractCommand
+{
+    public override void Execute(object? parameter)
+    {
+        var aboutDialog = new AboutDialog
+        {
+            Owner = Application.Current.MainWindow
+        };
+        aboutDialog.ShowDialog();
+    }
+}